🐛 fix: fix page crash when using browser as the stt engine (#1884)
* 🐛 fix: stt react hooks error

* 🐛 fix: warn active state in microphone button

* ♻️ refactor: move hooks

* ♻️ refactor: delete `export`

---------
